FIX: SD card code removal, code cleanup, iBus optimisation
The code for the SD card reader/writer/data-logger has been removed, as
the module did not work reliably. In case of error it hanged the whole
board and it stopped the telemetry. We could bypass this hang, but why
have a component that took 50% of the space on the memory and put it in
danger of data overflow, complicate the board soldering and add to the
weight of the final plane.

iBus sensor array is now better and more logical, so the most important
data is on the first few sensors. This part also need more testing to
assure the values are displayed correctly.