FINAL VERSION 1.0
Set option for changing between Serial com. over USB port - Serial
Monitor on the PC and Serial com. over RX/TX port of Arduino for
sending data to the receiver.

There are some optimisation options for the compass (calibration should
be performed for every sensor individually). Ours is mounted in the
wrong direction so there are some calculations necessary for it to
report correct heading.

Everything is tested in real life example and works as intended.